[ADD] 미니프로젝트-도서관리웹
1. SEQUENCE로 대여번호 입력
2. 대여관리 css 완성
3. 해야할 일: 팀원들과 코드 합치기

+++ src/main/java/com/ulgi/lease/controller/LeaseDeleteServlet.java
@@ -28,12 +28,18 @@ public class LeaseDeleteServlet extends HttpServlet {
 	 * @see HttpServlet#doGet(HttpServletRequest request, HttpServletResponse response)
 	 */
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		int leaseNo = Integer.parseInt(request.getParameter("leaseNo"));
-		int result = new LeaseService().deleteByLeaseNo(leaseNo);
-		if(result > 0) {
-			response.sendRedirect("/lease/list");
+
+		if(Boolean.parseBoolean(request.getParameter("del"))) {
+			int leaseNo = Integer.parseInt(request.getParameter("leaseNo"));
+			int result = new LeaseService().deleteByLeaseNo(leaseNo);
+			
+			if(result > 0) {
+				response.sendRedirect("/lease/list");
+			}else {
+				request.getRequestDispatcher("/WEB-INF/views/common/error.jsp").forward(request, response);
+			}
 		}else {
-			request.getRequestDispatcher("/WEB-INF/views/common/error.jsp").forward(request, response);
+			response.sendRedirect("/lease/list");
 		}
 	}
